requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
cloudscraper==1.2.71
Django==4.2.6
django-etc==1.4.0
django-hitcount==1.3.5
gunicorn==21.2.0
idna==3.4
packaging==23.2
psycopg==3.1.12
psycopg2==2.9.9
pyparsing==3.1.1
python-dotenv==1.0.0
requests==2.31.0
requests-toolbelt==1.0.0
soupsieve==2.5
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
whitenoise==6.6.0
